% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEuclideanDistanceMatrix}
\alias{getEuclideanDistanceMatrix}
\title{Get Euclidean Distance Matrix}
\usage{
getEuclideanDistanceMatrix(data)
}
\arguments{
\item{data}{A distance matrix (dist()) or a square matrix.}
}
\value{
A symmetric matrix where each element (i, j) represents the Euclidean distance between the i-th and j-th items.
}
\description{
This function computes the Euclidean distance matrix for the given data matrix.
}
\details{
The function calculates the Euclidean distance between each pair of rows in the input data matrix. 
The resulting distance matrix is symmetric, with the distance between item i and item j being the same as the distance between item j and item i.
}
\examples{
\dontrun{
data <- matrix(runif(100), nrow=10)
distance_matrix <- getEuclideanDistanceMatrix(data)
}

}
