% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHSLocalMdsStress}
\alias{getHSLocalMdsStress}
\title{Get Local MDS Stress}
\usage{
getHSLocalMdsStress(
  data,
  confVec,
  neighborhood,
  notNeighborhood,
  Rn,
  tt,
  Gamma
)
}
\arguments{
\item{data}{A distance matrix (dist()) or a square matrix.}

\item{confVec}{A numeric vector representing the configuration vector.}

\item{neighborhood}{A binary matrix indicating neighborhood relationships.}

\item{notNeighborhood}{A binary matrix indicating non-neighborhood relationships.}

\item{Rn}{An integer specifying the number of dimensions for the configuration.}

\item{tt}{A numeric value representing the weight for the non-neighborhood stress.}

\item{Gamma}{A numeric constant value used for smoothing the configuration.}
}
\value{
A numeric value representing the local MDS stress.
}
\description{
This function calculates the local Multidimensional Scaling (MDS) stress for the given data.
}
\details{
The function first reshapes the configuration vector into a matrix and computes the Euclidean distance matrix for the configuration. 
It then calculates the difference between the original data and the configuration distance matrix, squares these differences, and 
multiplies them by the neighborhood matrix. The non-neighborhood stress is calculated similarly but using the non-neighborhood matrix.
The final stress value is the difference between the neighborhood stress and the weighted non-neighborhood stress.
}
\examples{
\dontrun{
data <- matrix(runif(100), nrow=10)
confVec <- runif(30)
neighborhood <- matrix(sample(0:1, 100, replace=TRUE), nrow=10)
notNeighborhood <- matrix(sample(0:1, 100, replace=TRUE), nrow=10)
Rn <- 3
tt <- 0.5
stress <- getLocalMdsStress(data, confVec, neighborhood, notNeighborhood, Rn, tt)
}

}
