% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHSLocalMdsStressGradient}
\alias{getHSLocalMdsStressGradient}
\title{Get HS Local MDS Stress Gradient}
\usage{
getHSLocalMdsStressGradient(
  data,
  confVec,
  neighborhood,
  notNeighborhood,
  Rn,
  tt,
  Gamma
)
}
\arguments{
\item{data}{A distance matrix (dist()) or a square matrix.}

\item{confVec}{A numeric vector representing the configuration vector.}

\item{neighborhood}{A binary matrix indicating neighborhood relationships.}

\item{notNeighborhood}{A binary matrix indicating non-neighborhood relationships.}

\item{Rn}{An integer specifying the number of dimensions for the configuration.}

\item{tt}{A numeric value representing the weight for the non-neighborhood stress.}

\item{Gamma}{A numeric constant value used for smoothing the configuration.}
}
\value{
A numeric vector representing the gradient of the local MDS stress.
}
\description{
This function calculates the gradient of the local Multidimensional Scaling (MDS) stress for the given data.
}
\details{
The function reshapes the configuration vector into a matrix and computes the Euclidean distance matrix for the configuration. 
It then calculates the difference between the original data and the configuration distance matrix, normalizes these differences, 
and computes the gradient based on the neighborhood and non-neighborhood relationships. The final gradient is returned as a vector.
}
\examples{
\dontrun{
data <- matrix(runif(100), nrow=10)
confVec <- runif(30)
neighborhood <- matrix(sample(0:1, 100, replace=TRUE), nrow=10)
notNeighborhood <- matrix(sample(0:1, 100, replace=TRUE), nrow=10)
Rn <- 3
tt <- 0.5
gradient <- getHSLocalMdsStressGradient(data, confVec, neighborhood, notNeighborhood, Rn, tt)
}

}
