% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHSMdsStressGradient}
\alias{getHSMdsStressGradient}
\title{Get HS MDS Stress Gradient}
\usage{
getHSMdsStressGradient(data, confVec, Rn, Gamma)
}
\arguments{
\item{data}{A distance matrix (dist()) or a square matrix.}

\item{confVec}{A numeric vector representing the configuration vector.}

\item{Rn}{An integer specifying the number of dimensions for the configuration.}

\item{Gamma}{A numeric constant value used for smoothing the configuration.}
}
\value{
A numeric vector representing the gradient of the local MDS stress.
}
\description{
This function calculates the gradient of the HS Multidimensional Scaling (MDS) stress for the given data.
}
\details{
The function reshapes the configuration vector into a matrix and computes the Euclidean distance matrix for the configuration. 
It then calculates the difference between the original data and the configuration distance matrix, normalizes these differences, 
and computes the gradient. The final gradient is returned as a vector.
}
\examples{
\dontrun{
data <- matrix(runif(100), nrow=10)
confVec <- runif(30)
Rn <- 3
gradient <- getHSMdsStressGradient(data, confVec, Rn)
}

}
