% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNeighborhoodMatrix}
\alias{getNeighborhoodMatrix}
\title{Get Neighborhood Matrix}
\usage{
getNeighborhoodMatrix(data, k, symmetric)
}
\arguments{
\item{data}{A distance matrix (dist()) or a square matrix.}

\item{k}{An integer specifying the number of nearest neighbors to consider.}

\item{symmetric}{A boolean indicating whether the neighborhood matrix should be symmetric.}
}
\value{
A binary matrix where a value of 1 indicates that the corresponding items are neighbors.
}
\description{
This function computes a neighborhood matrix based on the input data matrix. 
The neighborhood matrix indicates which items are neighbors to each other.
}
\details{
The function first computes a ranked distance matrix from the input data. 
If `symmetric` is TRUE, the neighborhood relationship is made symmetric by considering the 
k-nearest neighbors for both items in the pair. If `symmetric` is FALSE, the neighborhood 
relationship is determined only based on the k-nearest neighbors of each item individually.
}
\examples{
\dontrun{
data <- matrix(runif(100), nrow=10)
k <- 3
symmetric <- TRUE
neighborhood_matrix <- getNeighborhoodMatrix(data, k, symmetric)
}

}
