% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNeighborhoodVector}
\alias{getNeighborhoodVector}
\title{Get Neighborhood Vector}
\usage{
getNeighborhoodVector(data, k)
}
\arguments{
\item{data}{A distance vector.}

\item{k}{An integer specifying the number of nearest neighbors to consider.}
}
\value{
A binary vector where a value of 1 indicates that the corresponding items are neighbors.
}
\description{
This function computes a neighborhood vector based on the input data vetor. 
The neighborhood vector indicates which items are neighbors to each other.
}
\details{
The function first computes a ranked distance vector from the input data. 
The neighborhood relationship is determined only based on the k-nearest neighbors of the 
vector individually.
}
\examples{
\dontrun{
data <- vector(runif(100), ncol=10)
k <- 3
symmetric <- TRUE
neighborhood_vector <- getNeighborhoodVector(data, k)
}

}
