% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getParameterT}
\alias{getParameterT}
\title{Get Parameter T}
\usage{
getParameterT(data, neighborhood, tau)
}
\arguments{
\item{data}{A distance matrix (dist()) or a square matrix.}

\item{neighborhood}{A binary matrix indicating neighborhood relationships.}

\item{tau}{A numeric value representing a scaling factor.}
}
\value{
A numeric value representing the parameter T.
}
\description{
This function calculates the parameter T, which represents repulsion, for the given data.
}
\details{
The function calculates the cardinality of the neighborhood and non-neighborhood relationships. 
It then computes the parameter T as the ratio of the neighborhood cardinality to the non-neighborhood cardinality, 
multiplied by the median of the data and the scaling factor tau.
}
\examples{
\dontrun{
data <- matrix(runif(100), nrow=10)
neighborhood <- matrix(sample(0:1, 100, replace=TRUE), nrow=10)
tau <- 0.5
parameterT <- getParameterT(data, neighborhood, tau)
}

}
