% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HSlocalMDS}
\alias{HSlocalMDS}
\title{Hyperbolic Smoothing Local MDS}
\usage{
HSlocalMDS(
  data,
  conf = NULL,
  Rn = 2L,
  Kproj = 5L,
  Kquality = 0L,
  verbose = FALSE,
  selectBetterUnitFree = TRUE,
  smallerUnitFree = 1e-04,
  n_t = 8L,
  ratio = 3.162278,
  applyHyperbolicSmoothing = TRUE,
  gamma = 1,
  n_gamma = 20L,
  rho = 0.3162278,
  maxIt = 100L,
  optMethod = "CG",
  optTrace = 0L,
  optReport = 10L
)
}
\arguments{
\item{data}{A distance matrix (dist()) or a square matrix.}

\item{conf}{An initial configuration. If none is supplied the configuration will be generated using PCA.}

\item{Rn}{The dimension of the space which the data will be represented in. The default is Rn=2.}

\item{Kproj}{The number of neighborhood to be preserved in the projection.}

\item{Kquality}{The number of neighborhood applied at Local Continuity to assess the quality of the projection. The default is NULL, set the same number at Kproj.}

\item{verbose}{A boolean constant. If true, information will be printed on terminal during the execution.}

\item{selectBetterUnitFree}{A boolean constant. If true, a Geometric sequence will be generated to select the best Unit Free.}

\item{smallerUnitFree}{the smallest value for the unit free parameter. Geometric sequence's start value.}

\item{n_t}{length for the grid search.}

\item{ratio}{ratio of the geometric sequence.}

\item{applyHyperbolicSmoothing}{A boolean constant. If true, the Hyperbolic Smoothing will be applyed.}

\item{gamma}{A numeric constant value used for smoothing the configuration.}

\item{n_gamma}{maximum number of gamma values allowed.}

\item{rho}{A numeric constant value between 0 and 1 used for decrease gamma value.}

\item{maxIt}{It is the control argument of the function optim.}

\item{optMethod}{The optimization method. The default is "CG", a conjugate gradients method.  it is possible to choose one of these methods: "Nelder-Mead", "BFGS", "CG", "L-BFGS-B".}

\item{optTrace}{Non-negative integer. If positive, tracing information on the progress of the optimization is produced. Higher values may produce more tracing information: for method "L-BFGS-B" there are six levels of tracing.}

\item{optReport}{The frequency of reports for the "BFGS", "L-BFGS-B" and "SANN" methods if optTrace is positive.}
}
\value{
Five components:

conf   The points projected in 'd' dimension space.

LocalContinuityResult   The criterion Local Continuity (LC).

Tau   The value of the parameter unit free that generate the best projection according to the criterion Local Continuity.

Tt   The repultion.

optimResult   A list with results from optimization process.
}
\description{
A Smoothing Optimization Approach Applied to the Local MDS Method
}
\examples{
# A U-shaped curve, (x^4+1), R^2 projected in R^1
library(NLDR)
set.seed(1)
x<-seq(0.9,1.5,0.05)
xx<-seq(-1,1,0.2)+runif(11,0,0.1)
x<-c(-x,x,xx)
Ccurve<-cbind(x,x^4+1)
d<-stats::dist (Ccurve)
Rn = 1

conf = as.matrix(cmdscale(d = d, k = Rn))

HSlocalMDSResult = HSlocalMDS(data = as.matrix(d), 
                                      conf = conf,
                                      Rn = Rn,
                                      Kproj = 5,
                                      Kquality = 5
)

titulo <- expression(paste("Blue lines connect the observation in ",
                            R^2," to the projection in ",R^1))
plot(x,x^4+1,ylim=c(-0,6),xlim=c(-6,6),asp=1, pch = 1,
     main="u-shaped curve projected to one dimensional space",
     sub= titulo,cex.sub=0.7,cex.main=0.7)
segments(x,x^4+1,HSlocalMDSResult$conf , rep(0,length(x)), col = "blue")
points(HSlocalMDSResult$conf,rep(0,length(x)), pch = 16,cex=0.7)
abline(h=0)

}
\references{
Lisha Chen and Andreas Buja, "Local multidimensional scaling for nonlinear dimension reduction, graph drawing, and proximity analysis." Journal of the American Statistical Association, 2009, 104(485), pp.209-219.
}
