% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RcppGetLocalContinuityMetaCriterion}
\alias{RcppGetLocalContinuityMetaCriterion}
\title{Local Continuity Meta Criterion}
\usage{
RcppGetLocalContinuityMetaCriterion(data, conf, Rn, k)
}
\arguments{
\item{data}{A distance matrix (dist()) or a square matrix.}

\item{conf}{The current configuration.}

\item{Rn}{The dimension of the space which the data will be represented in. The default is Rn=2.}

\item{k}{The number of neighborhood applied at Local Continuity to assess the quality of the projection.}
}
\value{
3 components:

Nk   Local Continuity Meta Criterion.

Mk   Local Continuity Meta Criterion Normalizado.

Mk_adjusted   Local Continuity Meta Criterion Normalizado Ajustado.
}
\description{
Local Continuity Meta Criterion
}
