% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RcppHSMDS}
\alias{RcppHSMDS}
\title{HS MDS}
\usage{
RcppHSMDS(
  data,
  conf,
  Rn = 2L,
  Kquality = 2L,
  verbose = FALSE,
  applyHiperbolicSmoothing = TRUE,
  gamma = 1,
  n_gamma = 30L,
  rho = 0.5,
  maxIt = 30L,
  optMethod = "CG"
)
}
\arguments{
\item{data}{A distance matrix (dist()) or a square matrix.}

\item{conf}{An initial configuration. If none is supplied the configuration will be randomly generated. If x=="cmdscale", the function cmdscale() is used to provide an inicial configuration.}

\item{Rn}{The dimension of the space which the data will be represented in. The default is Rn=2.}

\item{Kquality}{The number of neighborhood applied at Local Continuity to assess the quality of the projection. The default is NULL, set the same number at Kproj.}

\item{verbose}{A boolean constant. If true, informations will be printed on terminal during the execution.}

\item{applyHiperbolicSmoothing}{A boolean constant. If true, the Hiperbolic Smoothing will be applyed.}

\item{gamma}{A numeric constant value used for smoothing the configuration.}

\item{n_gamma}{maximum number of gamma values allowed.}

\item{rho}{A numeric constant value between 0 and 1 used for decrease gamma value.}

\item{maxIt}{It is the control argument of the function optim.}

\item{optMethod}{The optimization method. The default is "CG", a conjugate gradients method.  it is possible to choose one of these methods: "Nelder-Mead", "BFGS", "CG", "L-BFGS-B".}
}
\value{
Five components:

conf   The points projected in 'd' dimension space.

LocalContinuityResult   The criterion Local Continuity (LC).

optimResult   A list with results from optimization process.
}
\description{
A Smoothing Optimization Approach Applied to the MDS Method
}
\examples{
# A U-shaped curve, (x^4+1), R^2 projected in R^1
set.seed(1)
x<-seq(0.9,1.5,0.05)
xx<-seq(-1,1,0.2)+runif(11,0,0.1)
x<-c(-x,x,xx)
Ccurve<-cbind(x,x^4+1)
d<-stats::dist (Ccurve)
dataset<- as.matrix(d)
Rn = 1


conf = cmdscale(d = dataset, k = Rn)
conf = as.matrix(conf)

RcppHSMDSResult = RcppHSMDS(data = as.matrix(d), 
                                      conf = conf,
                                      Kquality = 5,
                                      Rn = 1,
                                      maxIt = 10000, 
                                      optMethod = "CG"
)

titulo <- expression(paste("Blue lines connect the observation in ",
                            R^2," to the projection in ",R^1))
plot(x,x^4+1,ylim=c(-0,6),xlim=c(-6,6),asp=1,
     main="u-shaped curve projected to one dimensional space",
     sub= titulo,cex.sub=0.7,cex.main=0.7)
segments(x,x^4+1,RcppHSMDSResult$conf , rep(0,length(x)), col = "blue")
abline(h=0)

}
