% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RcppOptimHSLocalMds}
\alias{RcppOptimHSLocalMds}
\title{Rcpp Optimization for Hyperbolic Smoothing Local MDS}
\usage{
RcppOptimHSLocalMds(
  data,
  conf,
  neighborhood,
  notNeighborhood,
  Rn,
  tt,
  Gamma,
  maxIt,
  optMethod
)
}
\arguments{
\item{data}{A distance matrix (dist()) or a square matrix.}

\item{conf}{An initial configuration.}

\item{neighborhood}{Symmetric matrix that represents which items are neighbors to each other.}

\item{notNeighborhood}{Symmetric matrix that represents which items are not neighbors to each other.}

\item{Rn}{The dimension of the space which the data will be represented in. The default is Rn=2.}

\item{tt}{The repultion.}

\item{Gamma}{A numeric constant value used for smoothing the configuration.}

\item{maxIt}{The number of optimizing iterations to assess the quality of the projection.}

\item{optMethod}{The optimization method. It is possible to choose one of these methods: "Nelder-Mead", "BFGS", "CG", "L-BFGS-B".}
}
\value{
optimResult   A list with results from optimization process.
}
\description{
Rcpp Optimization for Hyperbolic Smoothing Local MDS
}
