% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RcppOptimHSMds}
\alias{RcppOptimHSMds}
\title{Rcpp Optimization for HS MDS}
\usage{
RcppOptimHSMds(data, conf, Rn, Gamma, maxIt, optMethod)
}
\arguments{
\item{data}{A distance matrix (dist()) or a square matrix.}

\item{conf}{An initial configuration.}

\item{Rn}{The dimension of the space which the data will be represented in. The default is Rn=2.}

\item{Gamma}{A numeric constant value used for smoothing the configuration.}

\item{maxIt}{The number of optimizing iterations to assess the quality of the projection.}

\item{optMethod}{The optimization method. It is possible to choose one of these methods: "Nelder-Mead", "BFGS", "CG", "L-BFGS-B".}
}
\value{
optimResult   A list with results from optimization process.
}
\description{
Rcpp Optimization for HS MDS
}
